business_id は uint64である必要がなさそう.

+++ database/row/user_business.schema.go
@@ -10,7 +10,7 @@ import (
 type UserBusiness struct {
     ID         uint64    `ddl:"auto" json:"id,string"`
     UserID     uint64    `db:"user_id" json:"user_id"`
-    BusinessID uint64    `db:"business_id" json:"business_id"`
+    BusinessID uint32    `db:"business_id" json:"business_id"`
     Level      uint32    `db:"level" json:"level"`
     LastBuyAt  time.Time `db:"last_buy_at" json:"last_buy_at"`
 }
